+++ src/main/java/com/example/demo/controller/UsuarioController.java
@@ -8,10 +8,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/usuarios")
@@ -36,5 +33,11 @@ public class UsuarioController {
     ) {
         UsuarioDTO dto = service.usuarioId(id);
         return ResponseEntity.ok(dto);
+    }
+
+    @PatchMapping("{id}/inativar")
+    public ResponseEntity<Void> inativar(@PathVariable Long id){
+        service.inativar(id);
+        return ResponseEntity.noContent().build();
     }
 }